fix(pipeline): split store modules to satisfy 300-line gate
Extract models, GitHub apply handlers, and correlation helpers from
pipeline_store.py so CI file-length check passes.
