+++ src/factory/nats/pipeline_store.py
@@ -2,84 +2,36 @@
 
 from __future__ import annotations
 
-from dataclasses import dataclass, field
 from datetime import UTC, datetime, timedelta
 from pathlib import Path
-from typing import Literal
 
 from factory.nats.pipeline_cf_registry import pages_branch_matches, resolve_pages_project
+from factory.nats.pipeline_correlation import active_cf_pending_row
 from factory.nats.pipeline_db import PipelineDb, default_db_path
-
-PipelineStageStatus = Literal[
-    "pending",
-    "running",
-    "success",
-    "failure",
-    "skipped",
-    "unknown",
-    "n/a",
-]
-
-DEFAULT_REPO = "Roxabi/roxabi-factory"
-
-M1_DEPLOY_QUORUM = frozenset(
-    {
-        "factory-hub",
-        "factory-telegram",
-        "factory-discord",
-        "factory-dashboard",
-        "factory-clipool",
-    }
+from factory.nats.pipeline_github_apply import (
+    apply_check_run,
+    apply_pull_request,
+    apply_workflow_run,
+    github_repo,
+)
+from factory.nats.pipeline_models import (
+    DEFAULT_REPO,
+    M1_DEPLOY_QUORUM,
+    PipelineCheckRow,
+    PipelineRunRow,
+    PipelineStageStatus,
+    now_iso,
+    parse_iso,
 )
 
-
-@dataclass(slots=True)
-class PipelineCheckRow:
-    name: str
-    status: str
-    conclusion: str | None = None
-
-
-@dataclass(slots=True)
-class PipelineRunRow:
-    repo: str
-    pr_number: int
-    title: str
-    head_sha: str | None
-    head_ref: str | None
-    html_url: str | None
-    reviewed: bool
-    open: bool
-    ci_status: PipelineStageStatus
-    merge_status: PipelineStageStatus
-    publish_status: PipelineStageStatus
-    m1_deploy_status: PipelineStageStatus
-    cf_deploy_status: PipelineStageStatus
-    publish_sha: str | None
-    checks: list[PipelineCheckRow] = field(default_factory=list)
-    last_event_at: str | None = None
-    updated_at: str | None = None
-
-
-def _now_iso() -> str:
-    return datetime.now(tz=UTC).isoformat()
-
-
-def _worst_ci(checks: list[PipelineCheckRow]) -> PipelineStageStatus:
-    if not checks:
-        return "unknown"
-    conclusions = {c.conclusion or c.status for c in checks}
-    if "failure" in conclusions or "cancelled" in conclusions:
-        return "failure"
-    statuses = {c.status for c in checks}
-    if statuses & {"queued", "in_progress", "pending"}:
-        return "running"
-    if conclusions <= {"success", "skipped", "neutral"} or statuses <= {
-        "completed",
-        "success",
-    }:
-        return "success"
-    return "unknown"
+__all__ = [
+    "DEFAULT_REPO",
+    "M1_DEPLOY_QUORUM",
+    "PipelineCheckRow",
+    "PipelineRunRow",
+    "PipelineStageStatus",
+    "PipelineStore",
+]
 
 
 class PipelineStore:
@@ -112,7 +64,7 @@ class PipelineStore:
             return False
         self._processed.add(event_id)
         if self._db is not None:
-            self._db.record_processed(event_id, _now_iso())
+            self._db.record_processed(event_id, now_iso())
         elif len(self._processed) > 10_000:
             self._processed.clear()
         return True
@@ -129,7 +81,7 @@ class PipelineStore:
             if cutoff is None:
                 rows.append(row)
                 continue
-            ts = _parse_iso(row.last_event_at) or _parse_iso(row.updated_at)
+            ts = parse_iso(row.last_event_at) or parse_iso(row.updated_at)
             if ts is None or ts >= cutoff:
                 rows.append(row)
         rows.sort(
@@ -142,7 +94,7 @@ class PipelineStore:
         return self._runs.get((repo, pr_number))
 
     def _touch(self, row: PipelineRunRow) -> None:
-        now = _now_iso()
+        now = now_iso()
         row.last_event_at = now
         row.updated_at = now
         self._save_row(row)
@@ -210,124 +162,15 @@ class PipelineStore:
         if trace_id and not self.mark_processed(trace_id):
             return
         payload = payload or {}
-        repo = str(payload.get("repository") or DEFAULT_REPO)
+        repo = github_repo(payload)
         action = str(payload.get("action") or "")
 
         if kind.startswith("pull_request."):
-            self._apply_pull_request(repo, action, payload, kind)
+            apply_pull_request(self, repo, action, payload, kind)
         elif kind.startswith("check_run."):
-            self._apply_check_run(repo, payload)
+            apply_check_run(self, repo, payload)
         elif kind.startswith("workflow_run."):
-            self._apply_workflow_run(repo, payload)
-
-    def _apply_pull_request(
-        self, repo: str, action: str, payload: dict[str, object], kind: str
-    ) -> None:
-        pr = payload.get("pull_request")
-        if not isinstance(pr, dict) or pr.get("number") is None:
-            return
-        pr_number = int(pr["number"])
-        reviewed = _pr_has_reviewed_label(pr.get("labels"))
-        if kind == "pull_request.labeled" and isinstance(payload.get("label"), dict):
-            name = payload["label"].get("name")
-            if name == "reviewed":
-                reviewed = True
-        if kind == "pull_request.unlabeled" and isinstance(payload.get("label"), dict):
-            name = payload["label"].get("name")
-            if name == "reviewed":
-                reviewed = False
-
-        if action == "closed":
-            merged = bool(pr.get("merged"))
-            row = self._upsert_pr(
-                repo=repo,
-                pr_number=pr_number,
-                title=str(pr.get("title") or ""),
-                head_sha=_str_or_none(pr.get("head_sha")),
-                head_ref=_str_or_none(pr.get("head_ref")),
-                html_url=_str_or_none(pr.get("html_url")),
-                open_=False,
-            )
-            row.reviewed = reviewed or row.reviewed
-            row.merge_status = "success" if merged else "skipped"
-            if merged:
-                row.publish_status = "pending"
-                row.m1_deploy_status = "pending"
-                row.cf_deploy_status = "pending"
-            else:
-                row.publish_status = "n/a"
-                row.m1_deploy_status = "n/a"
-                row.cf_deploy_status = "n/a"
-            self._touch(row)
-            return
-
-        row = self._upsert_pr(
-            repo=repo,
-            pr_number=pr_number,
-            title=str(pr.get("title") or ""),
-            head_sha=_str_or_none(pr.get("head_sha")),
-            head_ref=_str_or_none(pr.get("head_ref")),
-            html_url=_str_or_none(pr.get("html_url")),
-            open_=True,
-        )
-        row.reviewed = reviewed or row.reviewed
-        if action == "synchronize" and row.head_sha:
-            row.ci_status = "running"
-        self._touch(row)
-
-    def _apply_check_run(self, repo: str, payload: dict[str, object]) -> None:
-        cr = payload.get("check_run")
-        if not isinstance(cr, dict):
-            return
-        numbers = cr.get("pull_request_numbers")
-        pr_numbers: list[int] = []
-        if isinstance(numbers, list):
-            pr_numbers = [int(n) for n in numbers if n is not None]
-        if not pr_numbers:
-            return
-        name = str(cr.get("name") or "check")
-        status = str(cr.get("status") or "unknown")
-        conclusion = _str_or_none(cr.get("conclusion"))
-        for pr_number in pr_numbers:
-            row = self._runs.get((repo, pr_number))
-            if row is None:
-                row = self._upsert_pr(repo=repo, pr_number=pr_number, open_=True)
-            replaced = False
-            for check in row.checks:
-                if check.name == name:
-                    check.status = status
-                    check.conclusion = conclusion
-                    replaced = True
-                    break
-            if not replaced:
-                row.checks.append(
-                    PipelineCheckRow(name=name, status=status, conclusion=conclusion)
-                )
-            row.ci_status = _worst_ci(row.checks)
-            self._touch(row)
-
-    def _apply_workflow_run(self, repo: str, payload: dict[str, object]) -> None:
-        wr = payload.get("workflow_run")
-        if not isinstance(wr, dict):
-            return
-        name = str(wr.get("name") or "")
-        if name != "publish":
-            return
-        conclusion = str(wr.get("conclusion") or "")
-        if conclusion not in {"success", "failure", "cancelled"}:
-            return
-        head_sha = _str_or_none(wr.get("head_sha"))
-        if conclusion == "success" and head_sha:
-            self._last_publish_sha = head_sha
-            if self._db is not None:
-                self._db.set_last_publish_sha(head_sha)
-        row = _active_publish_pending_row(self._runs.values(), repo)
-        if row is None:
-            return
-        row.publish_status = "success" if conclusion == "success" else "failure"
-        if head_sha:
-            row.publish_sha = head_sha
-        self._touch(row)
+            apply_workflow_run(self, repo, payload)
 
     def apply_cloudflare_event(
         self,
@@ -358,7 +201,7 @@ class PipelineStore:
         expected_branch = target["branch"]
         if not pages_branch_matches(pages_dict, expected_branch):
             return
-        row = _active_cf_pending_row(self._runs.values(), repo)
+        row = active_cf_pending_row(self._runs.values(), repo)
         if row is None:
             return
         row.cf_deploy_status = status
@@ -375,9 +218,8 @@ class PipelineStore:
             return
         if not kind.endswith("converge.completed"):
             return
-        now = _now_iso()
         if self._db is not None:
-            self._db.set_meta("last_converge_at", now)
+            self._db.set_meta("last_converge_at", now_iso())
 
     def recompute_m1_from_fleet(
         self,
@@ -406,58 +248,3 @@ class PipelineStore:
                     break
             row.m1_deploy_status = "success" if ok else "pending"
             self._touch(row)
-
-
-def _str_or_none(value: object) -> str | None:
-    return str(value) if isinstance(value, str) and value else None
-
-
-def _active_publish_pending_row(
-    runs: object,
-    repo: str,
-) -> PipelineRunRow | None:
-    candidates = [
-        row
-        for row in runs
-        if row.repo == repo
-        and not row.open
-        and row.merge_status == "success"
-        and row.publish_status == "pending"
-    ]
-    if not candidates:
-        return None
-    return max(candidates, key=lambda row: row.pr_number)
-
-
-def _active_cf_pending_row(runs: object, repo: str) -> PipelineRunRow | None:
-    candidates = [
-        row
-        for row in runs
-        if row.repo == repo
-        and not row.open
-        and row.merge_status == "success"
-        and row.cf_deploy_status == "pending"
-    ]
-    if not candidates:
-        return None
-    return max(candidates, key=lambda row: row.pr_number)
-
-
-def _parse_iso(value: str | None) -> datetime | None:
-    if not value:
-        return None
-    try:
-        return datetime.fromisoformat(value)
-    except ValueError:
-        return None
-
-
-def _pr_has_reviewed_label(labels: object) -> bool:
-    if not isinstance(labels, list):
-        return False
-    for item in labels:
-        if item == "reviewed":
-            return True
-        if isinstance(item, dict) and item.get("name") == "reviewed":
-            return True
-    return False
